refactor: Refactoring according to feedback
