+++ src/main/java/racingcar/domain/Game.java
@@ -1,9 +1,15 @@
 package racingcar.domain;
 
+import racingcar.view.EmptyNameException;
+import racingcar.view.NotProperInputTypeException;
+import racingcar.view.TooLongNameException;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class Game {
+    private final static int LONGNAME_STANDARD = 5;
+
     private List<Car> cars = new ArrayList<>();
     private int tryCount;
 
@@ -47,5 +53,26 @@ public class Game {
 
     public void setCar(List<Car> cars) {
         this.cars = cars;
+    }
+
+    public void catchNameException(String[] name) throws EmptyNameException, TooLongNameException {
+        for (String n : name) {
+            catchEmptyName(n);
+            catchTooLongName(n);
+        }
+    }
+
+    public void catchEmptyName(String name) throws EmptyNameException {
+        if (name == "") throw new EmptyNameException();
+    }
+
+    public void catchTooLongName(String name) throws TooLongNameException {
+        if (name.length() > LONGNAME_STANDARD) throw new TooLongNameException();
+    }
+
+    public void catchTryCountException(String count) throws NotProperInputTypeException {
+        boolean isDigit = count.matches("[+-]?\\d*(\\.\\d+)?");
+        if (!isDigit) throw new NotProperInputTypeException();
+
     }
 }
